Refactor: 유효성 체크 문구를 생성할 때 받아서 체크할 수 있게 수정
